+++ src/main/java/com/tamrielnetwork/gbooster/gbooster/listeners/JobsPayment.java
@@ -14,6 +14,6 @@ public class JobsPayment implements Listener {
 
     @EventHandler
     public void onPayment(JobsPaymentEvent event){
-        event.set(CurrencyType.MONEY, (event.getPayment().get(CurrencyType.MONEY)/(float)main.getActiveBoostersManager().getJobsBooster())*(float)main.getActiveBoostersManager().getBoosterMultiplier(BoosterType.JOBS_MONEY, true));
+        event.set(CurrencyType.MONEY, (event.getPayment().get(CurrencyType.MONEY)/main.getActiveBoostersManager().getJobsBooster())*main.getActiveBoostersManager().getBoosterMultiplier(BoosterType.JOBS_MONEY, true));
     }
 }
